+++ lib/providers/app_provider.dart
@@ -87,6 +87,19 @@ class AppProvider extends ChangeNotifier {
     notifyListeners();
   }
 
+  void updateMessageUnsentSatus(String messageId) async {
+    notifyListeners();
+    _messages = _messages.map((message) {
+      if (message.id == messageId) {
+        message.unsent = true;
+        return message;
+      }
+
+      return message;
+    }).toList();
+    notifyListeners();
+  }
+
   void updateSelectedTicketMessage(String? id) {
     _selectedTicketMessageId = id;
     notifyListeners();